rearrange redux index - user and resume
