+++ src/utilis/SideBar.tsx
@@ -1,19 +1,18 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import { portfolioReducer } from "../reducers/PortfolioContent";
-import { resumeReducer } from "../reducers/ResumeContent";
-import firebase from "./firebase";
-import { RootState } from "../reducers";
 import { useDispatch, useSelector } from "react-redux";
-import {
-  portfolioLoading,
-  resumeLoading,
-  setAlert,
-  userLoading,
-} from "../action";
+import { useNavigate } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faHeart } from "@fortawesome/free-solid-svg-icons";
-import { useNavigate } from "react-router-dom";
+
+import { portfolioReducer } from "../reducers/PortfolioContent";
+import { resumeReducer } from "../reducers/ResumeContent";
+import { RootState } from "../reducers";
+import { userLoading } from "../action/UserReducerAction";
+import { resumeLoading } from "../action/ResumeReducerAction";
+import { portfolioLoading, setAlert } from "../action";
+
+import firebase from "./firebase";
 
 const SideBarArea = styled.div`
   position: fixed;
